requirements.txt
pilmoji
numpy
pillow